@@ -1,5 +1,5 @@
 {
-    "name": "elnebuloso\/flex-converter",
+    "name": "elnebuloso/flex-converter",
     "type": "library",
     "license": "MIT",
     "description": "different converter, easy to use",
@@ -15,11 +15,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4.0"
     },
     "autoload": {
         "psr-0": {
-            "Flex\\": "src\/"
+            "Flex\\": "src/"
         }
     }
 }
