@@ -10,18 +10,17 @@ use Flex\Converter\ArrayToXml;
  */
 class ArrayToXmlTest extends \PHPUnit_Framework_TestCase
 {
-
     /**
      * @test
      */
     public function testCreateSimple()
     {
         $expected = '<?xml version="1.0" encoding="UTF-8"?><books><foo>1</foo><bar>2</bar><baz>3</baz></books>';
-        $books = array(
+        $books = [
             'foo' => 1,
             'bar' => 2,
-            'baz' => 3
-        );
+            'baz' => 3,
+        ];
 
         $converter = new ArrayToXml();
         $result = $this->filter($converter->convert('books', $books));
@@ -34,13 +33,13 @@ class ArrayToXmlTest extends \PHPUnit_Framework_TestCase
     public function testCreateForCollection()
     {
         $expected = '<?xml version="1.0" encoding="UTF-8"?><books><element>foo</element><element>bar</element><element>baz</element></books>';
-        $books = array(
-            'element' => array(
+        $books = [
+            'element' => [
                 'foo',
                 'bar',
-                'baz'
-            )
-        );
+                'baz',
+            ],
+        ];
 
         $converter = new ArrayToXml();
         $result = $this->filter($converter->convert('books', $books));
@@ -53,42 +52,42 @@ class ArrayToXmlTest extends \PHPUnit_Framework_TestCase
     public function testCreateFull()
     {
         $expected = '<?xml version="1.0" encoding="UTF-8"?><books type="fiction"><book author="George Orwell"><title>1984</title></book><book author="Isaac Asimov"><title><![CDATA[Foundation]]></title><price>$15.61</price></book><book author="Robert A Heinlein"><title><![CDATA[Stranger in a Strange Land]]></title><price discount="10%">$18.00</price></book></books>';
-        $books = array(
-            '@attributes' => array(
-                'type' => 'fiction'
-            ),
-            'book' => array(
-                array(
-                    '@attributes' => array(
-                        'author' => 'George Orwell'
-                    ),
-                    'title' => '1984'
-                ),
-                array(
-                    '@attributes' => array(
-                        'author' => 'Isaac Asimov'
-                    ),
-                    'title' => array(
-                        '@cdata' => 'Foundation'
-                    ),
-                    'price' => '$15.61'
-                ),
-                array(
-                    '@attributes' => array(
-                        'author' => 'Robert A Heinlein'
-                    ),
-                    'title' => array(
-                        '@cdata' => 'Stranger in a Strange Land'
-                    ),
-                    'price' => array(
-                        '@attributes' => array(
-                            'discount' => '10%'
-                        ),
-                        '@value' => '$18.00'
-                    )
-                )
-            )
-        );
+        $books = [
+            '@attributes' => [
+                'type' => 'fiction',
+            ],
+            'book'        => [
+                [
+                    '@attributes' => [
+                        'author' => 'George Orwell',
+                    ],
+                    'title'       => '1984',
+                ],
+                [
+                    '@attributes' => [
+                        'author' => 'Isaac Asimov',
+                    ],
+                    'title'       => [
+                        '@cdata' => 'Foundation',
+                    ],
+                    'price'       => '$15.61',
+                ],
+                [
+                    '@attributes' => [
+                        'author' => 'Robert A Heinlein',
+                    ],
+                    'title'       => [
+                        '@cdata' => 'Stranger in a Strange Land',
+                    ],
+                    'price'       => [
+                        '@attributes' => [
+                            'discount' => '10%',
+                        ],
+                        '@value'      => '$18.00',
+                    ],
+                ],
+            ],
+        ];
 
         $converter = new ArrayToXml();
         $result = $this->filter($converter->convert('books', $books));
@@ -146,18 +145,18 @@ class ArrayToXmlTest extends \PHPUnit_Framework_TestCase
     private function filter($string)
     {
         $string = preg_replace(
-            array(
+            [
                 '/ {2,}/',
                 '/<!--.*?-->|\t|(?:\r?\n[ \t]*)+/s',
                 '/>\s*/',
-                '/\s*</'
-            ),
-            array(
+                '/\s*</',
+            ],
+            [
                 ' ',
                 '',
                 '>',
-                '<'
-            ),
+                '<',
+            ],
             $string
         );
 
