@@ -10,7 +10,6 @@ use stdClass;
  */
 class ArrayToObject
 {
-
     /**
      * convert array to object
      *
@@ -21,7 +20,7 @@ class ArrayToObject
     {
         if (is_array($data)) {
             if (!is_numeric(key($data))) {
-                return (object) array_map(__METHOD__, $data);
+                return (object)array_map(__METHOD__, $data);
             }
 
             return $data;
